requirements.txt
langchain-openai==0.1.15
python-dotenv==1.0.1
langchain-core==0.2.13

